requirements.txt
torch
torchvision
librosa
numpy
scikit-learn
matplotlib
seaborn
pyyaml
torchaudio